+++ tests/test_metadata.py
@@ -4,7 +4,7 @@ from typing import List
 import pytest
 
 from openeo.metadata import CollectionMetadata, Band, SpatialDimension, Dimension, TemporalDimension, BandDimension, \
-    MetadataException
+    MetadataException, DimensionAlreadyExistsException
 
 
 def test_metadata_get():
@@ -478,6 +478,17 @@ def test_metadata_add_band_dimension():
     assert new.band_names == ["red"]
 
 
+def test_metadata_add_band_dimension_duplicate():
+    metadata = CollectionMetadata({
+        "cube:dimensions": {
+            "t": {"type": "temporal"}
+        }
+    })
+    metadata = metadata.add_dimension("layer", "red", "bands")
+    with pytest.raises(DimensionAlreadyExistsException, match="Dimension with name 'layer' already exists"):
+        _ = metadata.add_dimension("layer", "red", "bands")
+
+
 def test_metadata_add_temporal_dimension():
     metadata = CollectionMetadata({
         "cube:dimensions": {
@@ -493,6 +504,17 @@ def test_metadata_add_temporal_dimension():
     assert new.dimension_names() == ["x", "date"]
     assert new.temporal_dimension.name == "date"
     assert new.temporal_dimension.extent == ["2020-05-15", "2020-05-15"]
+
+
+def test_metadata_add_temporal_dimension_duplicate():
+    metadata = CollectionMetadata({
+        "cube:dimensions": {
+            "x": {"type": "spatial"}
+        }
+    })
+    metadata = metadata.add_dimension("date", "2020-05-15", "temporal")
+    with pytest.raises(DimensionAlreadyExistsException, match="Dimension with name 'date' already exists"):
+        _ = metadata.add_dimension("date", "2020-05-15", "temporal")
 
 
 def test_metadata_drop_dimension():